+++ sshd-core/src/main/java/org/apache/sshd/client/auth/pubkey/KeyPairIdentity.java
@@ -20,15 +20,17 @@ package org.apache.sshd.client.auth.pubkey;
 
 import java.security.KeyPair;
 import java.security.PublicKey;
-import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 
 import org.apache.sshd.common.NamedFactory;
-import org.apache.sshd.common.NamedResource;
 import org.apache.sshd.common.config.keys.KeyUtils;
 import org.apache.sshd.common.session.SessionContext;
 import org.apache.sshd.common.signature.Signature;
+import org.apache.sshd.common.signature.SignatureFactoriesHolder;
 import org.apache.sshd.common.signature.SignatureFactoriesManager;
+import org.apache.sshd.common.signature.SignatureFactory;
 import org.apache.sshd.common.util.ValidateUtils;
 
 /**
@@ -36,14 +38,15 @@ import org.apache.sshd.common.util.ValidateUtils;
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public class KeyPairIdentity implements PublicKeyIdentity {
-    private final KeyPair pair;
-    private final Collection<NamedFactory<Signature>> signatureFactories;
+public class KeyPairIdentity implements PublicKeyIdentity, SignatureFactoriesHolder {
+    protected final KeyPair pair;
+    private final List<NamedFactory<Signature>> signatureFactories;
 
     public KeyPairIdentity(SignatureFactoriesManager primary, SignatureFactoriesManager secondary, KeyPair pair) {
-        this.signatureFactories = ValidateUtils.checkNotNullAndNotEmpty(
-                SignatureFactoriesManager.resolveSignatureFactories(primary, secondary),
-                "No available signature factories");
+        this.signatureFactories = Collections.unmodifiableList(
+                ValidateUtils.checkNotNullAndNotEmpty(
+                        SignatureFactoriesManager.resolveSignatureFactories(primary, secondary),
+                        "No available signature factories"));
         this.pair = Objects.requireNonNull(pair, "No key pair");
     }
 
@@ -53,12 +56,17 @@ public class KeyPairIdentity implements PublicKeyIdentity {
     }
 
     @Override
+    public List<NamedFactory<Signature>> getSignatureFactories() {
+        return signatureFactories;
+    }
+
+    @Override
     public byte[] sign(SessionContext session, byte[] data) throws Exception {
         String keyType = KeyUtils.getKeyType(getPublicKey());
-        Signature verifier = ValidateUtils.checkNotNull(
-                NamedFactory.create(signatureFactories, keyType),
-                "No signer could be located for key type=%s",
-                keyType);
+        // SSHD-1104 check if the key type is aliased
+        NamedFactory<? extends Signature> factory = SignatureFactory.resolveSignatureFactory(keyType, getSignatureFactories());
+        Signature verifier = (factory == null) ? null : factory.create();
+        ValidateUtils.checkNotNull(verifier, "No signer could be located for key type=%s", keyType);
         verifier.initSigner(session, pair.getPrivate());
         verifier.update(session, data);
         return verifier.sign(session);
@@ -69,7 +77,7 @@ public class KeyPairIdentity implements PublicKeyIdentity {
         PublicKey pubKey = getPublicKey();
         return getClass().getSimpleName()
                + " type=" + KeyUtils.getKeyType(pubKey)
-               + ", factories=" + NamedResource.getNames(signatureFactories)
+               + ", factories=" + getSignatureFactoriesNameList()
                + ", fingerprint=" + KeyUtils.getFingerPrint(pubKey);
     }
 }